# packages/api/src/index.ts
export { createApiApp } from './app.js'
export {
  apiContractOperations,
  apiContractSchemas,
  generateOpenApiDocument,
  type ApiContractMethod,
  type ApiContractOperation,
  type ApiContractParameter,
} from './contract.js'
export {
  apiAuthenticationMiddleware,
  installApiTokenRoutes,
  requireApiScope,
  type ApiAuthentication,
  type ApiSessionResolver,
  type ApiTokenMetadata,
  type ApiTokenService,
  type AuthenticatedApiToken,
  type IssuedApiToken,
  type SessionPrincipal,
} from './auth.js'
export {
  installGeneralResourceRoutes,
  type GeneralResourceRouteOptions,
} from './general-resources.js'
export type {
  ApiContext,
  ApiInstaller,
  AppInstaller,
  CreateApiAppOptions,
  UserPrincipal,
  UserProfile,
} from './context.js'
export {
  ApiError,
  DEFAULT_MAX_JSON_BODY_BYTES,
  errorResponse,
  notFoundResponse,
  readJsonBody,
  validationError,
  type ApiErrorBody,
  type FieldError,
  type JsonBodyOptions,
} from './errors.js'
export {
  cursorPage,
  type CursorPageEnvelope,
  type CursorSource,
  type CursorWindow,
} from './pagination.js'
export {
  installEmailLogRoutes,
  type EmailLogReader,
} from './email-log.js'
export {
  installPasswordAuthRoutes,
  type AuthDelivery,
  type AuthMailer,
  type PasswordAuthRouteOptions,
  type PasswordAuthService,
  type PasswordSessionIssuer,
} from './password-auth.js'
export {
  installReportRoutes,
  serializeClientRollup,
  serializeProjectBudget,
  serializeUninvoiced,
  type ReportReader,
} from './reports.js'
export {
  assertValidOidcProviderConfig,
  installOidcRoutes,
  OIDC_STATE_COOKIE_NAME,
  OIDC_START_RATE_WINDOW_MS,
  OIDC_TRANSACTION_TTL_MS,
  OIDC_TRANSACTION_RETENTION_MS,
  type OidcClientAuthentication,
  type OidcIdentityResolver,
  type OidcProviderConfig,
  type OidcRouteOptions,
  type OidcSessionIssuer,
  type OidcTransaction,
  type OidcTransactionStorePort,
} from './oidc.js'
export * from './resources/index.js'
export { serializeMany, serializeOne, type Serializer } from './serializer.js'
export {
  createApiSessionService,
  installSessionRoutes,
  SESSION_COOKIE_NAME,
  type ApiSessionService,
  type SessionMetadata,
  type SessionRevocationReason,
  type SessionStorePort,
} from './sessions.js'